Ensure renderer binding is scoped to the application namespace
The binding passed to the renderer is not scoped to the application
namespace, and as such, all application classes must be prefixed with
the namespace module.

This change emulates the binding being created from within the
application namespace by delegating const lookups to said namespace if
the constant is not in scope from the Kangaru::Controller class.